fix: changes register plugin to back-off mechanism
There is no longer a max number of tries for registering the plugin with
the exchange module. Instead there is now a mechanism based on time,
starting off with every 10s, then every 1min, and lastly every 10min.

The drawback is that new registration messages (144/day) are put on the
queue for processing at a later point by the exchange module.

Refs: FART-593